+++ src/components/Footer.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
-const Footer = ({ timeout }) => (
-  <footer id="footer" style={timeout ? { display: 'none' } : {}}>
+const Footer = () => (
+  <footer id="footer">
     <p className="copyright">
       &copy; Thomas Dillard - Dimension. Design:{' '}
       <a href="https://html5up.net">HTML5 UP</a>. Built with:{' '}
@@ -10,9 +9,5 @@ const Footer = ({ timeout }) => (
     </p>
   </footer>
 );
-
-Footer.propTypes = {
-  timeout: PropTypes.bool,
-};
 
 export default Footer;
