+++ src/components/articles/index.js
@@ -1,0 +1,3 @@
+export { Contact } from './Contact';
+export { Projects } from './Projects';
+export { About } from './About';
